% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fozzie_difference_join_rs}
\alias{fozzie_difference_join_rs}
\title{Rust internal function performing difference joins}
\usage{
fozzie_difference_join_rs(
  df1,
  df2,
  by,
  how,
  max_distance,
  distance_col,
  nthread
)
}
\description{
Rust internal function performing difference joins
}
