% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fozzie_string_join_rs}
\alias{fozzie_string_join_rs}
\title{Rust internal function performing string distance joins}
\usage{
fozzie_string_join_rs(
  df1,
  df2,
  by,
  method,
  how,
  max_distance,
  distance_col,
  q,
  max_prefix,
  prefix_weight,
  nthread
)
}
\description{
Rust internal function performing string distance joins
}
