% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{normalize_by}
\alias{normalize_by}
\title{Join columns expect a named list, where names are left-hand columns to join
on, and values are right-hand columns to join on. This function ensures a
fuzzy-like syntax to the user while producing the correct output for the
rust join utilities.}
\usage{
normalize_by(df1, df2, by)
}
\description{
Join columns expect a named list, where names are left-hand columns to join
on, and values are right-hand columns to join on. This function ensures a
fuzzy-like syntax to the user while producing the correct output for the
rust join utilities.
}
