% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fozziejoin-package.R
\docType{package}
\name{fozziejoin-package}
\alias{fozziejoin}
\alias{fozziejoin-package}
\title{fozziejoin: Performant Data Frame Joins with Inexact String Matching}
\description{
Dataframes can be joined based on inexact matching using Levenshtein edit distance and similar algorithms.
}
\author{
\strong{Maintainer}: Jon Downs \email{jon@jondowns.net}

}
\keyword{internal}
