% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fozziejoin-package.R
\docType{package}
\name{fozziejoin-package}
\alias{fozziejoin}
\alias{fozziejoin-package}
\title{fozziejoin: Efficient Data Frame Joins with Inexact Matching}
\description{
data frames can be joined based on inexact matching using common string distance and other fuzzy join methodologies.
}
\author{
\strong{Maintainer}: Jon Downs \email{jon@jondowns.net}

}
\keyword{internal}
